// wwwsrc/app/hema/business/models/reference/IObjectType.ts
export interface IObjectType {
   category: string;
   applianceType: string;
   applianceTypeDescription: string;
   applianceSafetyNotRequiredIndicator: string;
   applianceCategory: string;
   useIaci: string;
   fetchGCCode: string;
   association: string;
   associationNumber: number;
   allowCreateInField: string;
   allowDeleteInField: string;
}